\name{get_treat}
\alias{get_treat}
\title{Generates number of treatments in a Design}

\description{
The get_treat function finds the number of treaments in a given design.
}

\usage{
get_treat(design)
}

\arguments{
\item{design}{A design in matrix format}
}

\details{
This function reads a design (D) in matrix format and finds the number of treatments (v) in the given design where rows are treated as blocks.
}

\value{
A list of treatments in a design.
}

\author{
1) Sayantani Karmakar
2) Md. Ashraful Haque
3) Cini Varghese
4) Seema Jaggi
5) Eldho Varghese
6) Mohd. Harun
}
