\name{rmv_list}
\alias{rmv_list}
\title{Removes a item from a list}

\description{
The rmv_list function removes the specified item from the given list.
}

\usage{
rmv_list(list, val)
}
\arguments{
\item{list}{A list of values}

\item{val}{A specific value}
}

\details{
This function reads a list of values and removes the specified value from that list.
}


\value{
It returns a list.
}

\author{
1) Sayantani Karmakar
2) Md. Ashraful Haque
3) Cini Varghese
4) Seema Jaggi
5) Eldho Varghese
6) Mohd. Harun
}
